Properly handle IPX socket descriptors used in select() writefds.
If any IPX sockets are passed to select() in the writefds set, substitute them
for the private UDP socket before calling the real select(), if it remains in
writefds upon return, remove it and restore the previously removed IPX sockets.

In short: Allow detecting when an IPX socket is ready to send.