Block certain calls momentarily when connecting to a DOSBox server.
Normally calls like bind() will immediately succeed or fail depending
on what address was requested and what addresses are assigned to the
machine, but when using a DOSBox IPX server we don't know our address
until it gets assigned to us, so block functions impacted by that to
avoid spurious errors when starting up.

If the server is down or slow then the calls will fail after a few
seconds rather than hanging indefinitely.