Add send rate limiting options.
These options allow capping the packet send rate in terms of bandwidth
or number of packets, when enabled, reaching/exceeding the rate limit
will make the thread sleep before returning from the send call, akin to
a frame rate limit.

This is useful for games which otherwise saturate the network and do
not handle packet loss or latency gracefully.